Clean up fact_sales
No joins are necessary for fact_sales. Since the surrogate keys are not dynamic (e.g. identity column type) and use the dbt_utils.surrogate_key, these joins are unnecessary. Traditionally, you would need to join to dims to get the surrogate keys, but this newer methodology has a number of advantages, this being one.